requirements.txt
Flask==1.1.4
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.5.1
Flask-Script==2.0.6
dropbox==11.35.0
Pillow==9.2.0
SQLAlchemy==1.4.42
Werkzeug==1.0.1
asttokens==2.0.8
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
contourpy==1.0.5
executing==1.1.1
gevent==22.10.2
gunicorn==20.1.0
ipython==8.5.0
itsdangerous==1.1.0
line-bot-sdk==2.3.0
matplotlib==3.6.0
markupsafe==2.0.1
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
protobuf==3.20.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyQt5==5.12.3
PyQt5-sip==12.11.0
python-dotenv==0.21.0
setuptools==49.2.1

